(waybridge) numerous x11-wayland fixes
The -standalone mode for arcan_xwm should now be working better,
tested against Xarcan using it there as the window manager.

A big race condition was also fixed with the pairing and event
queue management for the compositor-surfaces that are actually
tied to x11 surfaces.

The principal issue was that surface-id can appear before x11-id,
but x11-id can also arrive (though rarely) before surface-id. Both
these pairing scenarios need to be handled. With the x11-id case
though, events on the x11-id pairing that needs to be forward to
the arcan window manager were being omitted instead of queued and
forwarded when the pairing was completed.

There are issues with input focus, life cycles etc. but this was
a big step forward.

FossilOrigin-Name: fecce804f091b2734eac31f5ba39fde23dadf0f8821b8c0bc647473637b21cc8