(shmif) add arcan_shmif_unlink()
When the shared resources are mapped, there is currently a problem to
hand-over a segment to another process after it has been acquired. This
prevents certain use cases, primarily handover for fsrv_encode/remoting
to arcan-net as well as monitoring tools. This change splits the unlink
stage into two, and allows the unlinking to be deferred via the
SHMIF_DONT_UNLINK flag as well as providing a manual unlink function
and a function to retrieve the current named key.