(net) rework argument parsing, shutdown command
A lot of the arcan-net argument parsing gets reworked. The meat of the
commit though is that fallback- migration and exit- redirect should be
working, with some minor kinks.

A missed detail still is that a migrate request remote-side will be
directed to the proxy (of course) and not the client, and it is not
certain how to resolve that because on the one side - it is desired
behavior, "move the proxied window to this other local connection
point", while at the same time it could be useful to continue hopping
networks.

A compromise would be some kind of detection that the new display server
is actually a network proxy, but in the end we likely need a URL scheme
to deal with this issue.

FossilOrigin-Name: 25485c4a19959902ed43273c8cd818abd3a73ab8bbe9ec235a3e148ac3ae54dd