(arcterm/tui) the big refactor
This practically rips up most of tui, moves things around, and splits
the unwieldy mess it has evolved into over the years. The big end goal
is to regain server-side rendering, with working bidi/shaping/...
which is still a bit out of reach.

This is a feature-regression in that the shaped/scrolling/... rendering
are temporarily dropped. Their performance and behavior before was
anything but stellar regardless, and never the default setting so.

Other minor nuissances is that the blink management and cursor drawing
styles were also stripped a bit, but will return to normal once the
server-side raster stage move has been completed.

The terminal event-loop is also rebuilt to be multithreaded by default.
This will cause it to be somewhat snapper and use less idle CPU, but
with the downside of possibly more synch- artifcats with slow
clients/shell.

On the plus side, it should be much clearer to navigate and work with
the code base, and the big gain is the first decoupling of the screen
drawing into a packed wire format.

There are some breaking details with the API for bindings and so on,
in order to make the core simpler and more features as 'opt-in'. The
ugliest of ducklings in all of this is the tsm_ use still, but it is on
its way out. Check the tui.h header with functions marked for
deprecation (not using the directive at the moment to not mess with
build systems too much).

Basically, the idea is to have the core functions deal with the 'screen'
as volatile/drawing in the same way a normal framebuffer would be across
resizes, making 'altscreen' the new normal.

The 'line-based' mode should be moved into an opt-in widget that deals
with the whole scrolling, line-wrapping etc. as there are so many more
language- aware features to take into account today versus when this was
only used for terminals.

Third Time's the charm.