(core) fastpath recordtargets
This should cut down on the superflous rendertarget updates when the
rendertarget is only being used to feed a frameserver (recordtarget).
It is still a short step behind the more efficient method of letting the
conductor stage drive the recordtarget update when both fences (vready
in frameserver and rendertarget update) and for exporting a
buffer-handle and let the recipient do synch- map.