(core/shmif/tui) nearer server-side text raster
This extends shmif with the buffer tag needed to change the format
to be interpreted as the TPACK (text-pack in liu of imagination).

Displayhint and fonthint calls now gets tracked and mirrored in the
frameserver context, so that they then can be mapped to a tui_raster
context accordingly.

The missing boilerplate and build-system modifications to reuse the
raster code on server side has also been added.

The rough outline for the attack plan is as follows:

1. get server-side raster working with correct font behavior
   - this includes removing the INC_ DEC_ font labels as they
     should be treated via normal fonthints

2. get font management synched with arcan_ttf so that glyph cache
   sharing works between same fonts of the same size

3. split out the cursor drawing so it is not done in raster,
   should be a simple vobj- like submit on agp_ activate
   - re-add cursor styles

4. update arcan-net shmif-server stage to correctly forward the
   video buffer

5. re-introduce shaped rendering

6. switch to atlas- styled backing, maintained by arcan_ttf as
   part of the normal font cache.