(core/shmif/tui) more server-side text
with TUI_RPACK env set, client buffers will be rasterised server side.
A number of rough corners still, but the basic principle and grunt work
is done.

This covers most of point 1. in the list in (5c8b859a), the font
management is desynched and forced to bitmap, and arcan-net needs
an update still.