(core) fallback to builtin- bitmap font
A long-standing wart has been that of the initial font state. The
default has been to require the scripts to set a font on startup,
or in the case of lwa, set the default font to the one provided by
the display server connection (if one was set).

This led to problems in 'quick and dirty' scripts where the age old
'fonts/default.ttf' convention was used, especially for testing
purposes. With the refactoring of the TUI font rendering to be done
server side, we now have a bitmap font at multiple sizes ready from
the start.

This patch changes the font state to revert to this builtin bitmap
font if the font subsystem is reverted/set to a missing font.

FossilOrigin-Name: b86353fe8c0cdac4979e4927cb0150e9f22b9a39e833ff447969dc48bafb0b07