(tui/core) rpack raster font-size
This makes the bitmap raster respect the density of the last sent
fonthint. The question remains open if that should prompt a direct
resize/reraster, or just calculate cell size and forward that to
the client. Both approaches has merit, visually the response will
be faster in the former case but since it is such a fringe operation
it might just be better the other way.

FossilOrigin-Name: 7ae60bd66f97013324133c189be6357b06bee57013729503a6f4ee98ae91deaf