(platform/video) add platform_video_decay
This is something of an ugly quirk/workaround for an issue that has
been prominant on intel video devices as of late, and likely ties into
glitches or 'laggy' input.

This will increase cpu utilization in cases where there are time gaps
between updates (so entirely event-driven input scenarios), but shouldn't
be noticeable elsewhere.

The real fix takes deep dives into the continued work on the conductor,
particularly the use of sync fences to ensure that we don't try to render
or scanout with an FBO that is still not being complete.

The workaround is handled by adding the concept of 'decay', from the
meaning of impulse -> sustain -> decay. Basically each platform gets to
report of the dirty- state of the rendering pipeline should be respected
or not.

This allows the egl-dri platform to say that after a dirty update has
arrived, the following n frames will be treated as dirty regardless,
making possibly double- triple- or quad- buffered scanout pipelines all
have the same contents.

FossilOrigin-Name: 27d05ddb6d00a7efcf680d3f73d900aa16bdb678aef907ce55a31d4000cfc3ee