(shmif/tui) dirty region debugging safeguards
This should fix the last round of visual anomalies when it comes
to the terminal emulator / tui clients post refactor. The main cause
was that the announced dirty region was out of bounds by one character,
which had the shmif_signal function early out.

A notable behavior change here is that on a successful video signal,
the dirty region will be reset to an invalid state. Clients are also
encouraged to use shmif_dirty calls instead of manually setting the
fields. This is to allow a later data packing mode where dirty regions
gets streamed into the vidp block.

As part of this, for debug builds, the controlling environment
variable of 'ARCAN_SHMIF_DIRTY_DEBUG' can be used to mess with video
signalling by flashing the dirty region green for a frame + delay,
then send the real one.