(core) resampleimage should respect alpha
This one was subtle and came with the changes to rendertarget clearcolor
as well as the type of the rendertarget being created. If the caller
used a shader that conveyed values in the alpha channel, they would be
masked out against the clearcolor.

FossilOrigin-Name: 38abcfc9bf626468349e281e42c546602416e8d2eacffc42bc1ed4a08c6c2bb0