(core) tui rpack size calculation
This should close the book on the size part for the pixel font server-
side rendering. The issue was that a. size calculations were just not
performed, and b. the hint updates from lua to frameserver state block
didn't acknowledge partial updates.

The last few parts of this equation (for now) then is to add the normal
font handling and caching, and shrink the shmif segment size during negotiation
to reap the first benefits. There is more to do, of course, when it comes
to sharing atlases.

FossilOrigin-Name: 5eed62b7cea9e6b67926e1bc25088a7d6c6ca9bc59d230e71e9291698d0c18a5