(tui) listwnd widget input label rework
This reduces the set of default labels for normal tui clients to
be depending on the internal state between the 'line based' vs the
'screen based' mode of operation. Ideally the line based approach
would be removed (which would allow the removal of the tsm screen
layer as well) and re-added as a widget layer, and this is a small
step towards that.

The list window now exposes labels for basic navigation, similar
to how the buffer window works. Some ugliness still missing in terms
of pagination, and other input details around scroll- hints.