(waybridge) enter/leave cleanup
Trying to narrow down the suspect input focus / gtk- instability
problems. This takes some related fixes in durden etc. as the
displayhints doesn't arrive properly. Other possible culprit is
that we never actually enter surfaces on an output (politely
omitting comments about that part of the protocol as well).

i#