(shmif/net) draft of CONNPATH=a12:// support
This is a bit big but shouldn't be that complicated. Basically, shmif is
extended to try and setup arcan-net work a preset socketpair, which
should make things a bit more ergonomic on the client side.

Mostly untested and need some extra polish still, but setting:

ARCAN_CONNPATH=a12://some.host:port /path/to/client

Should be able to forward a single client if arcan-net is installed and
present in the normal exec() search paths.