(waybridge) regression on output-enter
The surface_enter event was sent to the wrong resource, and too often.
Revert to the semi-old behavior, which also is "so-so", the better
solution would be to use the visibility bit in the displayhint, but
more important wayland issues to hunt for now.

FossilOrigin-Name: cde8db32831705914bee0f82c22b04518206e7dd0e881487fe6218a43db9169c