(lua) let open_nonblock connect to dgram
This changes some of the connection logic for when the path to setup
already exists. Previously this was assumped to be FIFO-in, but now
will check both socket state, allow RDRW and handle both stream and
dgram (in AF_UNIX). Still won't allow namespaces outside appl-temp,
as such management should go through afsrv_*.