(arcterm+security) add exec, keep_alive arguments
A long standing control with the terminal frameserver has been that there
was no real non-interactive method of controlling what gets executed inside
the terminal. The workaround for using the terminal to execute shell commands
and grabbing the results etc. hsa been to add the command explicitly through
env and arcan_db and launch_target into that. There was also an even uglier
workaround by manually injecting keypresses. This was so cumbersome and
limiting, while the protection against command injection in the scripting
domain marginal enough that it is really only useful in specialized context.

Therefore, this patch adds an exec and keep_alive argument that are reachable
from the normal launch_avfeed Lua call appended arguments. Thus:

    launch_avfeed("exec=ls -l:keep_alive", "terminal",
    function(source, status)
    end);

Would provide the window with the results of executing ls.