(core/lua) retain dirty- region on SUBREGION
Client updates can (currently) be done in two ways, one by simply
invalidating the current frame and replacing it with a new one, or
setting a partial region that gets updated. For the second mode
(SUBREGION) the region information was discarded after being sent to the
GPU.

This patch extends the 'frame' events delivered on a buffer update if
the client is in verbose reporting mode (target_flags) to also contain
region information if available and accurate.