(shmif/lua) extend content hint with constraints
Previously there was no way for a client to define what it's desired
resize constraints were, the option was merely to set the desired size
and then react to displayhint events or asking when requesting a new
subsegment. To better assist window layout and resizing, the content
hints that otherwise indicate scrolling states has been extended to
allow for min/max w/h.