(core/shmif) extend decode, fix rpack/vobj calc.
A number of large shmif fixes are included in this one, and requires
bumping the shmif version. First the _decode frameserver is split up
and cleaned in order to allow for more types. Then the rpack size
calculation is fixed to pick only the needed size.

This should bring the memory usage of a tui client down significantly.
The last big optimization we can do in this regard is to get rid of
truetype(and png and zlib that comes with them) and have a simplified
font management build that uses a stripped down version of the stb
TTF parser that only provides glyph index lookup (no ligatures etc.)
and use harfbuzz for the main one.

That optimization is best of for when the server- side properly binpacks
and deals with fonts though.