(tui/debugif) add flag for mouse control
The removal of the _defaults setup structure also removed the control
for the implementations to say if the mouse cursor input is useful or
not. While this interface might need to be rethought again when server
side text processing is the default, it is not something that can be
fixed at the moment.

FossilOrigin-Name: 127540ffbbb3ff4b7c1571f0869323bd8464095a201fa6aada3f730b53595bd9