(shmif/lua) add forced-fallback segment push
This extends the target_alloc behavior with the ability to force
the default behavior, i.e. the allocation won't reach client code
but be consumed by the IPC library.

       target_alloc(vid, cb, "debug", true)

Will force the built-in debug-if to be triggered immediately and
not only if the client fails to provide one.

FossilOrigin-Name: fb14dbffe4ba6131c64fd9d8106864e08e709808d268a6a4920ce0045ff0234a