(debugif) add new-window option to intercept
This adds handling for requesting new windows and binding the exception
control to that as a new thread. While useful in its own right, it also
paves the way for the missing hard part of allowing bidirectional socket
interception without dealing with split-UI-split-interception in the
same window (which would take a rewrite of bufferwnd etc. to get
anywhere).