(tui/shmif/lua) viewport/embedding/handover
This cleans up the management for handover execution and lets either the
new process or the old process define how it embeds into the surface of
the other. The previous handover_ api definition changes somewhat
(though in a backward compatible way) to allow controls over what
happens to stdio- files. The alloc functions that spawn a segment also
returns the cookie as part of the function returns.