(shmif) lower guardthread sleep interval
The guard-thread is based on checking the aliveness of a set pid at a
fixed interval. Ideally this should be the fallback and use other
primitives for detecting connection death or DMS pull.

On BSDs we can use kqueues and possibly fdfork- like tracking, and there
is a pidfd like mechanism coming in more recent linuxes. This quick
change lowers the interval from 5 to 1 seconds between tries. The
detection is still the biggest limiter to recovery speeds, but this
should at least make it slightly more tolerable.