(vr/lua) send hmd-prjection as metadata
After some experimenting back and forth the solution now is to
communicate the projection matrices from the HMD and let the script
decide if to use the normal camtag based projection or the HMD supplied
one.

This also extends the camtag_model function to have an arugment
form where the projection matrix can be exchanged for a custom one.