(waybridge) work on re-adding shm conversion
This should be mostly plugged in, though likely that we need to have
a buffer rotation of the normal [3] size (display, in flight, drawing).

The remaining blocker is that, for some reason, EGL refuses to create
an EGLImage based on the display being marked as invalid.

FossilOrigin-Name: 2297b6fec199eed14a966283c2f04eb2e1fe1965493352abf33e05c38611def2