(shmif) add helper for fenv-, bump version
This is mainly to help in extreme edge cases, such as the shm repacking
stage of waybridge where a proper shmifext- gl environment is present.

It adds the option of extracting the agp fenv, the set of gl functions
needed to implement the platform/agp_platform rendering functions.

This includes the texture upload and download code, which is what
waybridge needs access to.