(waybridge) continued hunt on the 'realize-stall'
So there is something subtle missing from the window manager that
prevents most, not all, clients from passing the xwl_realize_window
and ensure_surface_for_window stages that are needed to pair surface
between X and Wayland. It can be hacked around by 'raising' the
window with wmctrl, and the pre2 tagged version does not have the
issue.