(waybridge) workaround for pairing issue
This one is not pretty and doesn't fix the issue present in the last few
commits entirely. What happens deep inside Xwayland is that the message
saying the name of the wayland surface and the x-surface gets lost deep
inside the X event routing mess.

This hits us more often than in other wayland compositors due to our
multiprocess- setup messing with timing and ordering, as the underlying
cause is some global-state+race condition terror (happens with xterm,
not with chromium for reference), but it seems likely that it is a cause
of many 'disappearing surface' class of problems everywhere.

The tactic, awaiting rootless mode for Xarcan where we don't need this
side-channel mess, is to tacitly assume that late- X events (e.g. map)
on a surface should be paired to the last unknown 'non-shell' bound
surface.