(core) aggressively free resources in vt-switch
This one was not entirely trivial, and some side effects may still
linger when it comes to more advanced objects (cubemaps, ...).
Basically the refactor in privilege-separation caused drmSet/DropMaster
to fail to hand over properly. This was easy enough as a fix in its own,
but that brought to like the part that the file descriptor of the
underlying card-node would change. The quick hack to this would be to
dup in place so that gbm- would be none the wiser, but that was just too
unsightly.

Instead we rip of the band-aid and completely tear down the card and
its displays on VT switch, and rebuild the context on startup. The
good bits to this is that we come much further in terms of being able to
both switch between GPUs and do non-cooperative multi-GPU as this is
proof that all the graphics states actually can be rebuilt on the same
or another GPU.