(decode) add proto(text, t2s, probe)
This adds support for more decode capabilities in the default decode
frameserver. The ones added this round is 'text' (which simply maps to
tui-bufferwnd), t2s (which uses espeak for text to speech conversion)
and probe (which uses libmagic). This, if enabled, should greatly
increase the preview and accessibility options in WMs.

The next planned steps for this is finishing the 3d model loading,
image loading (just lift from aloadimage) and, maybe, an 'auto' selector
that picks the best path based on detected magic type.