(hookscripts) add a hookscript to enable dbgstall
This introduces a delay to the spawn of every frameserver, giving
more time to be able to attach a debugger before the thing spawns.

It is likely that the debugstall feature will be disabled later
in order to simply use shmif-debugif directly, or a version where
a special value to the debugstall will trigger it.

Problem with that is that normal launch_xxx functions do not allow
any subsegments (such as debug) so those functions need to be able
to be intercepted and bound to something that makes sense in the
appl.

This requires a convention in the .lua layer for registering window
creation hook functions, which should possible be a thing anyhow to
allow many more hook-scripts that offer wm agnostic features.