(build/fsrv-net) drop fsrv-net implementation
The previous net implementation was mostly an experiment to grok the Lua
API boundary and requirements. For the 0.6 series, this will be remade
to actually deal with the directory- server/ discovery layer, likely by
settling on one of the interesting 'future net' like solutions for this
specific task, and then use the a12 implementation for the actual
communication and security.