(build) rework m, rt, pthreads linking
This shovels around some of the stronger CMake eyesores (rt, m,
pthreads) and bumps the version to >= 3.1.0. In the process these
libraries are using find modules that expose 'import targets'.

The old hijack libraries are no longer built as part of the
project as their usefullness is practically gone. Also drop
some of the initial hook script attempts (input playback/record).