(a12/shmif) boilerplate for compression-fwd
In the cases where a client has a compressed buffer, it is tragic if
that gets decompressed in the client only to get recompressed on the
server side (even though bitrates might vary).

The tactic here is to add another _sub structure to shmif which, if
allowed, lets the client forward compressed frames directly. This gives
us a negotiation path to only ever decode at the final destination.

This patch provides the first round of plumbing for the feature, along
with small changes to shmif-server in order to allow the ack/nak on the
substructure, as well as a way to register that a channel provides its
video frames raw rather than directly encode into shmif.

It is likely that we should rework this a bit and only let a12 work on
non-shmif contexts, and use the helpers to perform the last translation.

FossilOrigin-Name: 4ea3cae5e8b04ae3e133feebebfdc571242f9c94be577409a0d9aff9f4cb9d71