(core/lua) add _fatal_handover entrypoint
This adds another entrypoint, _fatal_handler, expected to return true if
it is deemed that the system is in a state where it can recover from a
scripting error. It is used in the cases where wrong arguments are
passed to native calls, or when native calls would otherwise land in an
arcan_fatal state.

Normally, scripts should not care about this, letting the system recover
by the normal reset->adopt setup is much less complicated in most cases,
and the _fatal entrypoint can already be used to save important
userdata.

The main use for this feature is to be able to use the _lua layer as a
proxy for an external party, primarly as a way of writing simpler
applications in other languages. For that case, we need this mechanic
along with a hookscript that performs the pack/upack/pcall.

Also drop dangling references to freeze_image.