(platform/egl-dri) defer KD_GRAPHICS to 1st fram
The previous behavior was to immediately set KD_GRAPHICS on the
controlling TTY in order to disable text echo and keyboard input.  This
had the problem in that any errors ocurring between init and first draw
, including appl parsing errors, would be missed unless STDOUT is
redirected to a file.

This patch changes that to perform the KD_GRAPHICS (on linux) when the
first modesetting called is applied as part of the first frame scanout.

Also plumb some more of the missing pieces for allocating FBOs on
scanout capable buffers when direct-scanout mode is possible.

FossilOrigin-Name: e2aa47043c713beddc5159d03dcc1bae85d5162f369b3e9899635272947f6f1f