(waybridge) move xcb processing to separate thread
This fixes the block on mapping issue. It seems like the core of it
all was a possible race or descriptor management issue when stepping
outside the normal xcb polling and switching to calling poll on our
own. By adding synch between the wm thread and the waybridge thread
surface not getting realized issue is done. This should bring the
whole xwl- saga to about a 60% point, remaining big ticket items are
the selection/dnd management along with the configure-resize races.