(agp) fix missing symbols in stub build
Reaching the point where it is time to support more agp backends,
primarily for network rendering - but with that software only builds and
possibly vulkan also make sense.

For this the plan is to flesh out the 'stub' agp platform so that
it can build frames of:

[shader updates] -> alloc, free, compile, ...
[storage updates] -> resize, upload, dirty, ...
[rendertarget updates] -> dst storage, projection, ...
[rendertarget draw-list] -> [shader + uniforms, vertices, txcos]

with an output packing format that can be streamed. Then create an
unpacker that works similarly (but calls into agp).