(core) wire up context robustness
This is the first step to the 'last' level of resilience tactics, being
able to recover from GPU failures. This simply checks for notifications
of a graphics context failure and performs a reset on the video platform.

It is, of course, hopefully naive and needs to be combined with stronger
measures - including GPU switching, revocation of GPU access to clients,
force-swap WM, simplifying / truncating shaders and, lastly, let the
psep parent 're-build/re-exec'.

The later part should also combine with a ping- stage between Lua VM
invocations and the other parts of the processing chain so that
live-locks of the 'while true do nothing' form can be attributed and
recovered.

FossilOrigin-Name: 3323b5a2b9854aa5444a44b4d5cb638351518dd4b8c457814808053230f263ea