(core) add ANR / livelock protection
This one has quite far-reaching consequences. The privilege-separated
part will now monitor a shared timestamp counter. If that counter stalls
for more than a fixed threshold of about 5 seconds, an interrupt signal
is sent.

This interrupt signal hooks into the lua state machine and generates a
scripting error, prompting the normal script error recover.  If that
does not resolve the situation, the child is forcibly killed and we move
into hard crash recovery.

Next iterations to consider for this tactic is to also have a SIGINT per
time-out restriction, and a chain-loader version used by the _lwa stage.