(terminal) switch to socketpair for inter-thread
This fixes a minor issue with the client provided debug interface
in that the creation and event come from the main thread, but it is
synched and populated in the terminal processing one. Switching the
pipe pair to a socket pair, multiplexing always in the processing
thread and signalling from the main thread gets the refresh correct.