(tui) proxy subwindow handler in widgets
Another method that was left out from proxying, this is likely going
to remain a problem due to the constraints of the design. Should
probably switch to a blacklist approach where the widget masks handlers
that shouldn't be processed, and proxy the rest by default rather than
the other way around.