(shmif/tui) change interface for _bgcopy
This fixes some of the stupidity of the previous one. Since it was
never released in any fashion, just fixing it instead of providing
a wrapper is a reasonable choice.

The main problem was that it didn't have any way of communicating
results or state, and the initial tests did not really have a need
for it. This adds a notification FD where the copy status will be
sent, either strictly 'finished or error' or (later) with periodic
progress report (flag field to control reporting).

The implementation is still a naive thread, but with this we can
switch to either kernel accelerated facilities or out-of-process
ones. The point is that it should just be a 'fire and forget' (or
asynch notify back) in order to properly respond to BCHUNK events,
and not having an interop implementation for that now would hurt
code complexity in the long run.

Updated the dnd test application to test the behavior.