(tui) mouse wheel to up/down (+shift to pg) in alt
If the mouse isn't in direct-to-client mode but rather handled by
the tui implementation, the wheel did not have much of an effect
when in altscreen (nothing tui 'knows' how to scroll).

Then it makes more sense to do something that might have a similar
effect, so now it is mapped to up-arrow, down-arrow (or with a shift
modifier, page-up, page-down).