(build/sdl) default to SDL2 for hybrid build
This adjusts two problems with the SDL use - one is that the default
platform for the everything- build was sdl1.2 rather than 2. The other
is that when no width/height was specified by the caller, it assumed 0x0
and some windowing systems didn't interpret that as 'whatever the WM
wants' but rather, ok, 0 sized window.

This also fixes a possible infinite loop in the initial flush for the
SDL1.2 platform, which seems to have hit X (only?).

A remaining issue is that the UTF8 values provided for key input is
still problematic, and the remaining workarounds probably need to be
solved in keyboard.lua - for sdl1.2 no inputs are received on recent X
(might be related to input driver) so all presses are empty. The other
is that for SDL2, the very concept is gone - you are expected to request
TEXTINPUT when you want that, but since we want the IME- like input on a
higher level, that is not a viable option. The current hack works ok on
OSX, but fails to accomodate shift states on X.

FossilOrigin-Name: 0cfed86abc887c13ea4d5bdfe99b360624892f4e46ffa40ce9bfe2e1569e30c4