(waybridge) prepare for clipboard-x11 handover
This makes some structural change that needs to be reflected in the WM.
First the old fallback of re-using control in -exec mode is droppped.
This will be used to grab a clipboard and inherit into the WM at startup
so any clipboard operations can synch there.  Second is that if the
bridge is created in authoritative mode and new connections can't
be spawned, the control will be used to launch new clients.