(core/tui) server-side rendering for truetype
This adds the second big batch of changes to get tui- clients to do all
their text rendering server-side.

f138c3b and related did the other heavy lifting of getting the bitmap
font part going, now fonts are being tracked as displayhint/fonthint
events are pushed to a frameserver, which was a big missing step.

The terminal (temporarily) accepts a "tpack" argument to enable the
output format, as it is not battle tested enough to be a safe default.

Now we are in a spot where it is useful to (in order):

0. share font cache with normal text rendering in engine.
1. add texture atlas support for resolved glyphs.
2. add a renderer that uses the atlas from 1 to slice out quads.
3. switch to using renderer from 2 by merging tpack buffers and
   using in the agp_vstore.
4. add a special surface type to the lua layer that allows an output
   grid which matches this format to complement render_text that still
   has to be rastered when VIDs are embedded.
5. re-add shaped rendering, handle LTR/RTL lines.
6. add shaped selection feedback to client.
7. add simplified font definitions that only carry glyph-indices and
   ligature substitutions.
8. re-add ligature support.
9. MSDFd texture atlas mode.
10. automatic input translation for tpack mode to cell-align input

Corner cases to look out for are image_access_storage like functions,
texture-coordinate based 'zoom' etc.