(wayland) xwl delete_window wrongly routed
First and foremost, the destroy-window command to ICCCM compliant
windows was not being sent to the right ID, causing the window to
stay alive.

Secondly, a shmif segment failing/getting destroyed upstream would
still remain in the pollset, causing the wayland protocol process
to spin.

There is still a 'not really solvable' difference of opinion in
lifecycle management in that a destroy in shmif is final - the client
doesn't get to popup a 'are you sure' dialog while ICCCM (and xdg)
allows the client to ignore any destroy, but always killing the
client would have it trigger even when it is on a non-essential
window, so damned if you do ...