(scripts/osdkbd) add initial on-screen keyboard
This is a somewhat weird and interesting one at the same time. It
works as both a standalone appl, i.e. :

    mkdir osdkbd
    ln -s arcan/data/scripts/builtin/osdkbd.lua osdkbd/osdkbd.lua
    arcan ./osdkbd

and as a support tool:

    system_load("builtin/osdkbd.lua")();
    kbd = osdkbd_build(... see bottom of source for example ...)
    kbd:show();

and as a hook script (more of a hack):

    arcan -H builtin/osdkbd.lua

It is still quite barebones and a lot of niceities missing in order
for it to be a truly competitive one (autocomplete, suggestions,
buffered mode, history, learning, fuzzing matching, external oracle,
drag+vector actions).