(scripts/mouse) add 'tap' event path
While this is strictly not a mouse handling event, using the same
hooks and code for providing touch-screen specific gestures is the
easiest interface to re-use with little impact on appls using it.

This adds a 'tap' event handler and a mouse_tap_at function that
doesn't affect other mouse state tracking. If the model works out,
it should probably be expanded with drag-from-to like behavior as
well.

FossilOrigin-Name: 785f455eb740c9752fdcaec0ace6c286e5792975507453bffd5585dbbd6c24c3