(arcterm) add initial cli mode and reset mgmt
This makes two changes, one is a 'cli' argument that switches execution
to use the readline widget to grab commands from the user rather than
using terminal emulation and a shell to do it. This will act as a
primitive for building own 'shells' both in-process and indirectly
through integration with a possible CLI in the WM.

As an example, a durden tool will be able to use this to integrate
with the menu HUD, and then have each executed command launch an arcan
app through the handover mechanism OR use a legacy mode for single
terminal commands into discrete windows.

The other change is related, and that is to let the 'exec=cmd' mode
of running afsrv_terminal handle RESET commands by re-executing the
same command and re-using the existing window.

The next path for this feature is something like:

(durden) -> afsrv_terminal(cli) -> types in command -> requests
subsegment -> handover- exec to afsrv_terminal with the subsegment
in cmd mode.