(lua/decode-t2s) rm load_movie, cleanup
load_movie has been deprecated in favor of launch_decode for a long
while, given how poor the _lua.c implementation of it was. Cleaned
it up somewhat and added the option of a nil resource argument.

This is useful for the recent _decode modes such as t2s. where a
resource is not needed.

FossilOrigin-Name: 6ef72b8437f9014ddf0329bcfc65e2facb02db6e48ffcd3ba101e711a83ed323