(a12) enable encryption
This signifies the first push with working encryption, so test but
do not trust.

What is mainly missing from this round is the x25519 exchange(s),
including rekeying - though most of the scaffolding is in place and
it is by far the easiest piece.

The support tool also needs store for keys and a lookup setup. The
documentation models how this is intended to work, but the
implementation is lacking.

A new test, core/a12loop has been added as a suite to eventually
cover all major transfer modes, configurations and buffering forms.