(arcterm-cli) added cd and mode builtin and prompt
This makes the cli-mode somewhat more integratable in durden and others,
though x11/wayland modes still not working, being dependent on changes
in the wayland improvements branch.

The cli gets two new commands, cd for changing the current directory
and 'mode' for setting the execution mode. The later determines if we
run through a normal handover, vt100 via afsrv_terminal or arcan-wayland.