(core/lua) add mask argument to reset_transform
Situations where you want to invalidate one transform but not another,
such as blend but not move which is common in scenarios where a
selection event 'blend' would collide with a move animation that gets
cancelled and replaced (window managers) - either requires you to
manually rebuild the correct transform, which imposes design limitations
to the client.

Adding a mask to only reset parts of a transformation chain helps with
this as then you can replace one but keep others intact.