(engine/lua) add mask to instant_image_transform
This adds a similar setup to instant_image_transform as was done in
(ad95930). The slight difference is that there is the awkward use of
bool- flags in the instant case that does not apply for reset.

Since we can't remove the bool flags and risk breaking APIs, we
type-differentiate between int and bool to get the mask behavior.

Getting both the 'trigger tags' and mask-instant behavior has so rare
uses that it is best avoiding the problem altogether and limit to
either-or.

FossilOrigin-Name: 7c1e8e9c6ba6c4c783e776bd850d584360867f3608639d2cf1d1752b10944ac4