(waybridge) confinement, callbacks and input
A number of changes to how popups are configured and gets their frame
callbacks triggered, this should help with some of the popup issues in
firefox. Refactored mouse input code to merge inputs on storms and
changed ordering between inputs and frame-callbacks so callbacks will
be triggered after input is flushed.