(core/lua) add tracing facility
For conductor profile tuning as well as debugging of more advanced
transfer methods, cooperative tracing is needed. This patch wires in
a script-controllable trace buffer that the engine as well as lua
scripts can jack into (via benchmark_tracedata).

Thereafter some key points of engine processing should be added.

The previous tracing facilities in arcan_lua.c should be left intact
and excluded from this as the specifics can be added from the lua
level if desired, and not futz with the build system any further.

Finally some support script that exports the format to the tracing
format that chrome and the likes can visualize, merged with similar
output from the client side of shmif.